Fix a couple of regressions introduced in TurboVNC 1.0:
RFB < 3.8 clients also expect rfbAuthOK to be sent by the server whenever standard VNC authentication is used.  Otherwise, the client (RealVNC 3.3.x, for instance) will appear to hang when connecting to the TurboVNC Server.

For clients that don't support the Tight security descriptor (RealVNC, TigerVNC, etc.), the TurboVNC Server was never advertising rfbAuthNone, even if that was the only auth method supported by the server.  Instead, it should advertise rfbAuthNone if it is the only auth method supported (but not otherwise.)


git-svn-id: svn://svn.code.sf.net/p/turbovnc/code/branches/1.2.x@2498 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b