Fixed several issues related to the display of dialogs.  In applet mode, the ownerless dialogs (New Connection, Password, SSH, etc.) disappear to the back of the window stack if you click away from them, and they can't be recovered with Alt-Tab.  Further, under Java 7 and Linux, that same thing happens in standalone mode.  Thus, we now set ownerless dialogs to be always on top.
Normally, if a JDialog is ownerless, it is assigned to a "shared frame", and the shared frame is registered with the task switcher, even though it has no taskbar icon.  Such is not the case with Java 7 and Linux.  You can make a JDialog truly ownerless (give it a null parent), which causes it to appear in the taskbar and the task switcher in all cases, but doing that requires Java 6 and later, so we can't do it as long as we're still supporting OS X 10.5 (which is limited to Java 5.)  Further, giving the JDialog a null parent makes it impossible to set its icon, since calling setIconImage() for a JDialog is only supported in later versions of Java (not Java 5.)

This patch also assigns the proper owner to the JDialog when displaying the Options or Clipboard dialogs.  This fixes various cosmetic issues, such as:
-- When the Options dialog was invoked from the New Connection dialog, it was not behaving entirely like a modal dialog.  It was possible to click on the New Connection dialog and bring it to the front.
-- In full-screen mode on Linux/Gnome, popping up the Options or Clipboard dialogs would cause the viewer to temporarily exit full-screen mode.


git-svn-id: svn://svn.code.sf.net/p/turbovnc/code/branches/1.2.x@2697 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b